py/requirements.txt
kafka-python==2.0.2
elasticsearch==7.17.4
scikit-learn==1.1.1
pandas==1.4.3
numpy==1.23.1
grpcio-tools==1.29.0